feat: add test helpers and refactor
