+++ pallets/account/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+libsecp256k1 = { workspace = true, optional = true }
 log = { workspace = true }
 parity-scale-codec = { workspace = true }
 scale-info = { workspace = true }
@@ -30,7 +31,8 @@ pallet-evm = { workspace = true }
 astar-primitives = { workspace = true }
 
 [dev-dependencies]
-hex = { workspace = true}
+ethers = { workspace = true }
+hex = { workspace = true }
 pallet-balances = { workspace = true }
 pallet-ethereum = { workspace = true }
 pallet-evm = { workspace = true }
@@ -41,6 +43,8 @@ default = ["std"]
 std = [
 	"hex/std",
 	"log/std",
+	"libsecp256k1",
+	"libsecp256k1/std",
 	"parity-scale-codec/std",
 	"scale-info/std",
 	"sp-std/std",
@@ -57,6 +61,7 @@ std = [
 	"pallet-ethereum/std",
 ]
 runtime-benchmarks = [
+	"libsecp256k1/hmac",
 	"frame-support/runtime-benchmarks",
 	"frame-system/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
